# services/trading-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing portfolios, executing trades, and handling financial market operations.",
        "title": "Trading Service API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/orders": {
            "get": {
                "description": "Returns a paginated and filtered list of orders. Clients see only their own orders, employees see all.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Get all orders",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (PENDING, APPROVED, DECLINED)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by direction (BUY, SELL)",
                        "name": "direction",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by completion status",
                        "name": "is_done",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ListOrdersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a buy or sell order for a listing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Create a new order",
                "parameters": [
                    {
                        "description": "Order details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.OrderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/orders/{id}/approve": {
            "patch": {
                "description": "Supervisor approves a pending order",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Approve a pending order",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.OrderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/orders/{id}/cancel": {
            "patch": {
                "description": "Cancel a pending or approved order that hasn't been fully executed",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Cancel an order",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.OrderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/orders/{id}/decline": {
            "patch": {
                "description": "Supervisor declines a pending order",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Decline a pending order",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.OrderResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.CreateOrderRequest": {
            "type": "object",
            "required": [
                "account_number",
                "direction",
                "listing_id",
                "order_type",
                "quantity"
            ],
            "properties": {
                "account_number": {
                    "type": "string"
                },
                "all_or_none": {
                    "type": "boolean"
                },
                "direction": {
                    "enum": [
                        "BUY",
                        "SELL"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.OrderDirection"
                        }
                    ]
                },
                "limit_value": {
                    "type": "number"
                },
                "listing_id": {
                    "type": "integer"
                },
                "margin": {
                    "type": "boolean"
                },
                "order_type": {
                    "enum": [
                        "MARKET",
                        "LIMIT",
                        "STOP",
                        "STOP_LIMIT"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.OrderType"
                        }
                    ]
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1
                },
                "stop_value": {
                    "type": "number"
                }
            }
        },
...
        "dto.ListOrdersResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.OrderSummaryResponse"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "dto.OrderResponse": {
            "type": "object",
            "properties": {
                "account_number": {
                    "type": "string"
                },
                "after_hours": {
                    "type": "boolean"
                },
                "all_or_none": {
                    "type": "boolean"
                },
                "approved_by": {
                    "type": "integer"
                },
                "contract_size": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "direction": {
                    "$ref": "#/definitions/model.OrderDirection"
                },
                "is_done": {
                    "type": "boolean"
                },
                "limit_value": {
                    "type": "number"
                },
                "listing_id": {
                    "type": "integer"
                },
                "listing_name": {
                    "type": "string"
                },
                "margin": {
                    "type": "boolean"
                },
                "order_id": {
                    "type": "integer"
                },
                "order_type": {
                    "$ref": "#/definitions/model.OrderType"
                },
                "price_per_unit": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer"
                },
                "remaining_portions": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/model.OrderStatus"
                },
                "stop_value": {
                    "type": "number"
                },
                "ticker": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "dto.OrderSummaryResponse": {
            "type": "object",
            "properties": {
                "contract_size": {
                    "type": "number"
                },
                "direction": {
                    "$ref": "#/definitions/model.OrderDirection"
                },
                "listing_name": {
                    "type": "string"
                },
                "order_id": {
                    "type": "integer"
                },
                "price_per_unit": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer"
                },
                "remaining_portions": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/model.OrderStatus"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "errors.AppError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "model.OrderDirection": {
            "type": "string",
            "enum": [
                "BUY",
                "SELL"
            ],
            "x-enum-varnames": [
                "OrderDirectionBuy",
                "OrderDirectionSell"
            ]
        },
        "model.OrderStatus": {
            "type": "string",
            "enum": [
                "PENDING",
                "APPROVED",
                "DECLINED"
            ],
            "x-enum-varnames": [
                "OrderStatusPending",
                "OrderStatusApproved",
                "OrderStatusDeclined"
            ]
        },
        "model.OrderType": {
            "type": "string",
            "enum": [
                "MARKET",
                "LIMIT",
                "STOP",
                "STOP_LIMIT"
            ],
            "x-enum-varnames": [
                "OrderTypeMarket",
                "OrderTypeLimit",
                "OrderTypeStop",
                "OrderTypeStopLimit"
            ]
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Authorization header using the Bearer scheme.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}